(refactor) remove bot package
