+++ bot/src/main/java/ru/tinkoff/edu/java/bot/CommandsHandler.java
@@ -1,10 +1,10 @@
-package ru.tinkoff.edu.java.bot.bot;
+package ru.tinkoff.edu.java.bot;
 
 import com.pengrad.telegrambot.model.Update;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import ru.tinkoff.edu.java.bot.bot.command.CommandHandler;
-import ru.tinkoff.edu.java.bot.bot.meta.State;
+import ru.tinkoff.edu.java.bot.command.CommandHandler;
+import ru.tinkoff.edu.java.bot.meta.State;
 import ru.tinkoff.edu.java.bot.dto.HandledUpdate;
 
 import java.util.List;
