+++ bot/src/main/java/ru/tinkoff/edu/java/bot/UpdatesProcessor.java
@@ -1,11 +1,11 @@
-package ru.tinkoff.edu.java.bot.bot;
+package ru.tinkoff.edu.java.bot;
 
 import com.pengrad.telegrambot.TelegramBot;
 import com.pengrad.telegrambot.UpdatesListener;
 import com.pengrad.telegrambot.model.Update;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import ru.tinkoff.edu.java.bot.bot.meta.State;
+import ru.tinkoff.edu.java.bot.meta.State;
 import ru.tinkoff.edu.java.bot.dto.HandledUpdate;
 
 import java.util.HashMap;
