+++ bot/src/main/java/ru/tinkoff/edu/java/bot/command/CommandHandler.java
@@ -1,10 +1,10 @@
-package ru.tinkoff.edu.java.bot.bot.command;
+package ru.tinkoff.edu.java.bot.command;
 
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.BaseRequest;
 import com.pengrad.telegrambot.response.BaseResponse;
-import ru.tinkoff.edu.java.bot.bot.meta.Command;
-import ru.tinkoff.edu.java.bot.bot.meta.State;
+import ru.tinkoff.edu.java.bot.meta.Command;
+import ru.tinkoff.edu.java.bot.meta.State;
 
 public interface CommandHandler<T extends BaseRequest<T, R>, R extends BaseResponse> {
     Command command();
