+++ bot/src/main/java/ru/tinkoff/edu/java/bot/command/StartCommandHandler.java
@@ -1,4 +1,4 @@
-package ru.tinkoff.edu.java.bot.bot.command;
+package ru.tinkoff.edu.java.bot.command;
 
 
 import com.pengrad.telegrambot.model.Update;
@@ -6,22 +6,18 @@ import com.pengrad.telegrambot.request.SendMessage;
 import com.pengrad.telegrambot.response.SendResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import ru.tinkoff.edu.java.bot.bot.meta.Command;
-import ru.tinkoff.edu.java.bot.bot.MessageSender;
-import ru.tinkoff.edu.java.bot.bot.meta.State;
-
-import java.util.Arrays;
-import java.util.stream.Collectors;
+import ru.tinkoff.edu.java.bot.meta.Command;
+import ru.tinkoff.edu.java.bot.MessageSender;
+import ru.tinkoff.edu.java.bot.meta.State;
 
 
 @Component
 @RequiredArgsConstructor
-public class HelpCommandHandler implements CommandHandler<SendMessage, SendResponse> {
+public class StartCommandHandler implements CommandHandler<SendMessage, SendResponse> {
     private final MessageSender messageSender;
-
     @Override
     public Command command() {
-        return Command.HELP;
+        return Command.START;
     }
 
     @Override
@@ -31,7 +27,6 @@ public class HelpCommandHandler implements CommandHandler<SendMessage, SendRespo
 
     @Override
     public SendMessage handle(Update update) {
-        return messageSender.send(update, Arrays.stream(Command.values())
-                .map(Command::toString).collect(Collectors.joining("\n")));
+        return messageSender.send(update, "Welcome!", true);
     }
 }
