+++ bot/src/main/java/ru/tinkoff/edu/java/bot/command/UntrackCommandHandler.java
@@ -1,4 +1,4 @@
-package ru.tinkoff.edu.java.bot.bot.command;
+package ru.tinkoff.edu.java.bot.command;
 
 
 import com.pengrad.telegrambot.model.Update;
@@ -6,24 +6,24 @@ import com.pengrad.telegrambot.request.SendMessage;
 import com.pengrad.telegrambot.response.SendResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import ru.tinkoff.edu.java.bot.bot.meta.Command;
-import ru.tinkoff.edu.java.bot.bot.MessageSender;
-import ru.tinkoff.edu.java.bot.bot.meta.State;
+import ru.tinkoff.edu.java.bot.meta.Command;
+import ru.tinkoff.edu.java.bot.MessageSender;
+import ru.tinkoff.edu.java.bot.meta.State;
 
 
 @Component
 @RequiredArgsConstructor
-public class TrackCommandHandler implements CommandHandler<SendMessage, SendResponse> {
+public class UntrackCommandHandler implements CommandHandler<SendMessage, SendResponse> {
     private final MessageSender messageSender;
 
     @Override
     public Command command() {
-        return Command.TRACK;
+        return Command.UNTRACK;
     }
 
     @Override
     public State state() {
-        return State.TRACK;
+        return State.UNTRACK;
     }
 
     @Override
